third_party/blink/web_tests/fast/url/invalid-urls-utf8-expected.txt

This ensure we do not mess up the bytes when an URL is partially parsed and is invalid.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL src should be http:///. Was http:.
FAIL src should be https:///. Was https:.
FAIL src should be ftp:///. Was ftp:.
PASS src is expected
FAIL src should be ws:///. Was ws:.
FAIL src should be wss:///. Was wss:.
FAIL src should be foo://tête à tête@host/. Was foo://t%C3%AAte %C3%A0 t%C3%AAte@host/.
FAIL src should be foo://user:tête à tête@host/. Was foo://user:t%C3%AAte %C3%A0 t%C3%AAte@host/.
PASS src is expected
PASS src is expected
FAIL src should be foo://user:password@[xxxxx]/tête à tête/. Was foo://user:password@[xxxxx]/t%C3%AAte %C3%A0 t%C3%AAte/.
FAIL src should be foo://user:password@[?tête à tête/. Was foo://user:password@[?t%C3%AAte%20%C3%A0%20t%C3%AAte/.
FAIL src should be foo://user:password@[?tête à tête]/. Was foo://user:password@[?t%C3%AAte%20%C3%A0%20t%C3%AAte]/.
FAIL src should be foo://user:password@host:tête à tête/. Was foo://user:password@host:t%C3%AAte %C3%A0 t%C3%AAte/.
FAIL src should be foo://user:password@host:80tête à tête. Was foo://user:password@host:80t%C3%AAte %C3%A0 t%C3%AAte.
FAIL src should be foo://user:password@host:tête à tête/. Was foo://user:password@host:t%C3%AAte %C3%A0 t%C3%AAte/.
FAIL src should be foo://user:password@host:80tête à tête. Was foo://user:password@host:80t%C3%AAte %C3%A0 t%C3%AAte.
FAIL src should be ws://host:tête. Was ws://host:t%C3%AAte/.
FAIL src should be ftp://host:tête. Was ftp://host:t%C3%AAte/.
FAIL src should be http://host:tête. Was http://host:t%C3%AAte/.
FAIL src should be https://host:tête. Was https://host:t%C3%AAte/.
PASS successfullyParsed is true

TEST COMPLETE
